Move Title out of the Header
- In about:newtab we want to display header but not the title.
- When assistant is displayed header buttons should be visible, but title should not be.
