+++ src/browser/Navigators/Navigator/Assistant.js
@@ -11,7 +11,6 @@ import * as Search from "./Assistant/search"
 import {StyleSheet, Style} from '../../../common/style';
 import {cursor} from '../../../common/cursor';
 import * as Unknown from '../../../common/unknown';
-import * as Layer from './Layer';
 
 /*::
 import type {Address, DOM} from "reflex";
@@ -243,7 +242,6 @@ const styleSheet = StyleSheet.create
       , position: 'absolute'
       , top: '0px'
       , width: '100%'
-      , zIndex: Layer.assistant
       }
     , expanded:
       { height: '100%'
