+++ src/browser/Navigators/Navigator/Header.js
@@ -3,24 +3,21 @@
 import {html, thunk, forward} from 'reflex';
 import * as Style from '../../../common/style';
 import {always} from '../../../common/prelude';
-import * as Title from './Header/Title';
+import * as Title from './Title';
 import * as ShowTabsButton from './Header/ShowTabsButton';
 import * as NewTabButton from './Header/NewTabButton';
 import * as BackButton from './Header/BackButton';
-import * as Layer from './Layer';
 
 /*::
 import type {Address, DOM} from "reflex"
 
 export type Model = string
 export type Action =
-  | { type: "EditInput" }
   | { type: "ShowTabs" }
   | { type: "OpenNewTab" }
   | { type: "GoBack" }
 */
 
-const tagTitle = always({ type: "EditInput" });
 const tagShowTabs = always({ type: "ShowTabs" });
 const tagNewTab = always({ type: "OpenNewTab" });
 const tagGoBack = always({ type: "GoBack" });
@@ -28,9 +25,7 @@ const tagGoBack = always({ type: "GoBack" });
 export const height = Title.outerHeight;
 
 export const render =
-  ( title/*:string*/
-  , secure/*:boolean*/
-  , canGoBack/*:boolean*/
+  ( canGoBack/*:boolean*/
   , address/*:Address<Action>*/
   )/*:DOM*/ =>
   html.header
@@ -40,11 +35,6 @@ export const render =
   , [ BackButton.view
       ( canGoBack
       , forward(address, tagGoBack)
-      )
-    , Title.view
-      ( title
-      , secure
-      , forward(address, tagTitle)
       )
     , NewTabButton.view
       ( forward(address, tagNewTab)
@@ -56,16 +46,12 @@ export const render =
   );
 
 export const view =
-  ( title/*:string*/
-  , secure/*:boolean*/
-  , canGoBack/*:boolean*/
+  ( canGoBack/*:boolean*/
   , address/*:Address<Action>*/
   )/*:DOM*/ =>
   thunk
   ( 'Browser/NavigatorDeck/Navigator/Header'
   , render
-  , title
-  , secure
   , canGoBack
   , address
   )
@@ -79,7 +65,6 @@ const styleSheet = Style.createSheet
       , height: `${height}px`
       , color: 'inherit'
       , background: 'inherit'
-      , zIndex: Layer.header
       }
     }
   )
