+++ src/browser/Navigators/Navigator/Header/BackButton.js
@@ -3,7 +3,7 @@
 import {html, thunk, forward} from 'reflex'
 import * as Style from '../../../../common/style'
 import {always} from '../../../../common/prelude'
-import * as Title from './Title'
+import * as Title from '../Title'
 
 /*::
 import type {Address, DOM} from "reflex"
