+++ src/browser/Navigators/Navigator/Progress/PolyfillView.js
@@ -2,7 +2,6 @@
 
 import {thunk, html} from 'reflex';
 import * as Style from '../../../../common/style';
-import * as Layer from '../Layer'
 
 /*::
 import * as Ref from '../../../../common/ref';
@@ -43,7 +42,6 @@ const styleSheet = Style.createSheet
       , overflow: "hidden"
 
       , position: 'absolute'
-      , zIndex: Layer.progress
       , top: '27px'
       , width: "100%"
       }
