+++ src/browser/Navigators/Navigator/Progress/ProgressView.js
@@ -2,7 +2,6 @@
 
 import {thunk, html} from 'reflex';
 import * as Style from '../../../../common/style';
-import * as Layer from '../Layer'
 
 /*::
 import * as Ref from '../../../../common/ref';
@@ -50,7 +49,6 @@ const styleSheet = Style.createSheet
       , backgroundSize: 0
 
       , position: 'absolute'
-      , zIndex: Layer.progress
       , top: '27px'
       , width: "100%"
       }
